// web/src/api/generated/models/ProjectInstructionInput.ts
/* This file was generated by Kubb. Do not edit manually. */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description ProjectInstructionInput schema
 */
export type ProjectInstructionInput = {
  /**
   * @type string | undefined
   */
  body?: string;
  /**
   * @type string | undefined
   */
  priority?: string;
  /**
   * @type string | undefined
   */
  title?: string;
};