// web/src/api/generated/models/ProjectInstructionListResponse.ts
/* This file was generated by Kubb. Do not edit manually. */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ProjectInstructionOutput } from "./ProjectInstructionOutput.ts";

/**
 * @description ProjectInstructionListResponse schema
 */
export type ProjectInstructionListResponse = {
  /**
   * @type array,null
   */
  instructions: ProjectInstructionOutput[] | null;
};