// web/src/api/generated/models/ProjectInstructionOutput.ts
/* This file was generated by Kubb. Do not edit manually. */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type ProjectInstructionOutput = {
  /**
   * @type string
   */
  body: string;
  /**
   * @type string
   */
  created_at: string;
  /**
   * @type boolean
   */
  enabled: boolean;
  /**
   * @type integer, int64
   */
  id: number;
  /**
   * @type string
   */
  priority: string;
  /**
   * @type integer, int64
   */
  project_id: number;
  /**
   * @type string
   */
  scope: string;
  /**
   * @type string
   */
  source: string;
  /**
   * @type string
   */
  title: string;
  /**
   * @type string
   */
  updated_at: string;
};